Rename module compare_method -> compare_impl_item
